Allow callback for missing translations
Rails makes it possible to raise an exception for a missing translation,
which can be useful in development and test environments to spot
undefined translations. This change adds an optional user-definable callback
that can be invoked any time a translation is missing. It could be used
in tests to identify undefined translations or in production
environments to report missing translations to an exception tracking
service.

When defined, the missing translation callback does not replace the
existing guess/missing translation options, it's invoked in addition to
the existing placeholders.

Close #477.